% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{chetah}
\alias{chetah}
\title{The CHETAH output of the subset of the Melanoma data (\code{\link{data_mel}})
, run with the Head-Neck reference (\code{\link{reference_hn}})}
\format{see \code{\link{CHETAHclassifier}}}
\source{
for the original data: \href{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE72056}{GEO}
}
\usage{
chetah
}
\description{
The CHETAH output of the subset of the Melanoma data (\code{\link{data_mel}})
, run with the Head-Neck reference (\code{\link{reference_hn}})
}
\references{
Tirosh et al. (2016) Science 6282:189-196
}
\keyword{datasets}
