% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_mel}
\alias{data_mel}
\title{A subset of the Melanoma data, from Tirosh et al. (2016), Science (\code{\link{data_mel}})}
\format{This is a sparse Matrix, with the cells in the columns and the genes in the rows}
\source{
for the original data: \href{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE72056}{GEO}
}
\usage{
data_mel
}
\description{
A subset of the Melanoma data, from Tirosh et al. (2016), Science (\code{\link{data_mel}})
}
\references{
Tirosh et al. (2016) Science 6282:189-196
}
\keyword{datasets}
