% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{input_mel}
\alias{input_mel}
\title{A SingleCellExperiment on which CHEATHclassifier is run using the \code{headneck_ref}
It holds subset of the Melanoma data, from Tirosh et al. (2016), Science (\code{\link{data_mel}})}
\format{This is a SingleCellExperiment}
\source{
for the original data: \href{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE72056}{GEO}
}
\usage{
input_mel
}
\description{
A SingleCellExperiment on which CHEATHclassifier is run using the \code{headneck_ref}
It holds subset of the Melanoma data, from Tirosh et al. (2016), Science (\code{\link{data_mel}})
}
\references{
Tirosh et al. (2016) Science 6282:189-196
}
\keyword{datasets}
