% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils_CHETAH.R
\name{PlotBox}
\alias{PlotBox}
\title{Plot boxplots grouped by a class variable}
\usage{
PlotBox(toplot, class, col = NULL, grad_col = NULL, limits = NULL,
  return = FALSE)
}
\arguments{
\item{toplot}{a numeric variable. When provided as a dataframe,
the colname will be used a a title}

\item{class}{a factor of the length of \code{toplot}
to indicate the groups}

\item{col}{a character vector of colors, with the
names of the factor levels}

\item{grad_col}{if defined, the jittered points of
the individual values of each cell behind
the boxplot will be colored by these gradient colors
based on their value}

\item{limits}{the limits of \code{toplot}}

\item{return}{instead of printing, return the ggplot object}
}
\value{
A ggplot object
}
\description{
Plot boxplots grouped by a class variable
}
\examples{
CD8 <- as.matrix(data_mel['CD8A', ])
types <- chetah$classification
PlotBox(toplot = CD8, class = types)
}
