% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{reference_hn}
\alias{reference_hn}
\title{A subset of the Head-Neck data from Puram et al. (2017) Cancer Cell.}
\format{A list of expression matrices. Each object is named as the cell type of the cells in that matrix.
Each matrix has the cell (names) in the colums and the genes in the rows.}
\source{
for the original data: \href{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE103322}{GEO}
}
\usage{
reference_hn
}
\description{
A subset of the Head-Neck data from Puram et al. (2017) Cancer Cell.
}
\references{
Puram et al. (2017) Cancer Cell 171:1611-1624
}
\keyword{datasets}
