% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tsne_mel}
\alias{tsne_mel}
\title{The result of running t-SNE on the \code{\link{data_mel}} (subset of Melanoma data).
This is matrix of two columns (the two dimensions), with the cells (names) in the rows}
\format{A matirx with the t-SNE coordinates of the melanoma data in 2 columns (for the two dimensions)}
\source{
for the original data: \href{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE72056}{GEO}
}
\usage{
tsne_mel
}
\description{
The result of running t-SNE on the \code{\link{data_mel}} (subset of Melanoma data).
This is matrix of two columns (the two dimensions), with the cells (names) in the rows
}
\references{
Tirosh et al. (2016) Science 6282:189-196
}
\keyword{datasets}
