% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils_CHETAH.R
\name{CHETAHshiny}
\alias{CHETAHshiny}
\title{Launch a web page to interactively go trough the classification}
\usage{
CHETAHshiny(input, redD = NA, input_c = NA)
}
\arguments{
\item{input}{a SingleCellExperiment on which \code{\link{CHETAHclassifier}} has been run}

\item{redD}{the name of the reducedDim of the input to use for plotting}

\item{input_c}{the name of the assay of the input to use.
\code{NA} (default) will use the first one.}
}
\value{
Opens a web page in your default browser
}
\description{
Launch a web page to interactively go trough the classification
}
\examples{
\dontrun{
CHETAHshiny(input = input_mel)
}
}
