% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils_CHETAH.R
\name{Classify}
\alias{Classify}
\title{(Re)classify after running \code{\link{CHETAHclassifier}} using a confidence threshold \cr
NOTE: In case of bulk reference profiles: only the correlations will be used,
as the data does not allow for profile or confidence scores to be calculated.}
\usage{
Classify(chetah, thresh)
}
\arguments{
\item{chetah}{the CHETAH object}

\item{thresh}{a confidence threshold between -0 and 2. \cr
Selecting 0 will classify all cells, whereas 2 will result i
n (almost) no cells to be classified. \cr
\emph{recommended}: between 0.1 (fairly confident) and 1 (very confident)}
}
\value{
a charachter vector of the cell types with the names of the cells
}
\description{
(Re)classify after running \code{\link{CHETAHclassifier}} using a confidence threshold \cr
NOTE: In case of bulk reference profiles: only the correlations will be used,
as the data does not allow for profile or confidence scores to be calculated.
}
\examples{
## Classify all cells
chetah$classification <- Classify(chetah, 0)

## Classify only cells with a very high confidence
chetah$classifcation <- Classify(chetah, 1)
}
